+++ packages/testing/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @clerk/testing
 
+## 1.7.0
+
+### Minor Changes
+
+- Adding `waitForSubscribeButton` and `confirmAndContinue` methods to checkout page object. ([#5913](https://github.com/clerk/javascript/pull/5913)) by [@panteliselef](https://github.com/panteliselef)
+
+### Patch Changes
+
+- Updated dependencies [[`1ff6d6e`](https://github.com/clerk/javascript/commit/1ff6d6efbe838b3f7f6977b2b5215c2cafd715f6), [`fbf3cf4`](https://github.com/clerk/javascript/commit/fbf3cf4916469c4e118870bf12efca2d0f77d9d8)]:
+  - @clerk/shared@3.9.0
+  - @clerk/types@4.58.1
+  - @clerk/backend@1.32.1
+
 ## 1.6.7
 
 ### Patch Changes